api/openai/gpt-image-2/requirements.txt
pydantic >= 2.0.0 # required for inference.sh
inferencesh # required for inference.sh
openai >= 1.0.0
httpx >= 0.25.0
Pillow >= 10.0.0
